+++ src/timeseriesflattener/specs/__init__.py
@@ -1,0 +1,10 @@
+from __future__ import annotations
+import datetime as dt
+
+
+def _lookdistance_to_timedelta(
+    lookdistance: float | tuple[float, float],
+) -> tuple[dt.timedelta, dt.timedelta]:
+    if isinstance(lookdistance, tuple):
+        return (dt.timedelta(days=lookdistance[0]), dt.timedelta(days=lookdistance[1]))
+    return (dt.timedelta(days=0), dt.timedelta(days=lookdistance))
